Moved LoggerAction to Shared map
